lmb-892: remove the impersonation menu and just show the AuPill in the application template
