--- app/components/impersonation-menu.hbs
@@ -1,31 +1,0 @@
-<div class="au-u-flex au-u-flex--vertical-center">
-  <AuPill @icon="user">{{this.adminLabel}}</AuPill>
-
-  {{!-- <AuDropdown
-    @title="Simulatie opties"
-    @hideText={{true}}
-    @alignment="right"
-    @icon={{this.icon}}
-    class="au-u-margin-left-tiny"
-    role="menu"
-  >
-    {{#if this.isImpersonating}}
-      <AuLink @route="impersonate" @icon="switch" role="menuitem">
-        Simuleer een andere bestuurseenheid
-      </AuLink>
-
-      <AuButton
-        @skin="link"
-        @icon="logout"
-        role="menuitem"
-        {{on "click" this.stopImpersonation}}
-      >
-        Simulatie stopzetten
-      </AuButton>
-    {{else}}
-      <AuLink @route="impersonate" @icon="login" role="menuitem">
-        Simuleer een bestuurseenheid
-      </AuLink>
-    {{/if}}
-  </AuDropdown> --}}
-</div>
