+++ app/controllers/application.js
@@ -58,7 +58,7 @@ export default class ApplicationController extends Controller {
     let group;
     let classification;
 
-    if (this.impersonation.isImpersonating) {
+    if (this.isImpersonating) {
       user = this.impersonation.originalAccount.gebruiker;
       group = this.impersonation.originalGroup;
       classification = group.belongsTo('classificatie').value();
@@ -160,5 +160,21 @@ export default class ApplicationController extends Controller {
       this.environmentInfo.title !== '' &&
       this.environmentName !== '{{ENVIRONMENT_NAME}}'
     );
+  }
+
+  get user() {
+    if (!this.isImpersonating) {
+      return null;
+    }
+
+    return this.currentSession.user;
+  }
+
+  get isImpersonating() {
+    return this.impersonation.isImpersonating;
+  }
+
+  get adminLabel() {
+    return this.isImpersonating ? `Admin: ${this.user.fullName}` : 'Admin';
   }
 }
